+++ tests/testthat/test-extension-dispatch.R
@@ -52,3 +52,32 @@ test_that("getEncounter honours rates_funcs for base objects", {
     expect_equal(getEncounter(params, t = 2), params@initial_n * 0 + 2,
                  ignore_attr = TRUE)
 })
+
+test_that("classless extensions do not trigger project dispatch", {
+    resetMizerSession()
+    withr::defer(resetMizerSession())
+
+    chain <- c(stats = "0.0")
+    registerExtensions(chain)
+
+    params <- NS_params
+    params@extensions <- chain
+    params@rates_funcs$Rates <- "classless_rates_for_dispatch_test"
+    params@rates_funcs$Encounter <- "classless_encounter_for_dispatch_test"
+
+    assign("classless_rates_for_dispatch_test",
+           function(params, n, n_pp, n_other, t = 0, effort, rates_fns, ...) {
+        list(classless = TRUE)
+    }, envir = .GlobalEnv)
+    assign("classless_encounter_for_dispatch_test",
+           function(params, n, n_pp, n_other, t = 0, ...) {
+        params@initial_n * 0 + t
+    }, envir = .GlobalEnv)
+    withr::defer(rm("classless_rates_for_dispatch_test",
+                    "classless_encounter_for_dispatch_test",
+                    envir = .GlobalEnv))
+
+    expect_identical(getRates(params)$classless, TRUE)
+    expect_equal(getEncounter(params, t = 3), params@initial_n * 0 + 3,
+                 ignore_attr = TRUE)
+})